+++ docs/conf.py
@@ -58,7 +58,7 @@ author = "Scientific Software"
 # the built documents.
 #
 # The short X.Y version.
-version = "0.27.0"
+version = "0.28.0"
 # The full version, including alpha/beta/rc tags.
 release = version