+++ src/zocalo/__init__.py
@@ -6,7 +6,7 @@ import logging
 
 __author__ = "Diamond Light Source - Data Analysis Group"
 __email__ = "dataanalysis@diamond.ac.uk"
-__version__ = "0.27.0"
+__version__ = "0.28.0"
 
 logging.getLogger("zocalo").addHandler(logging.NullHandler())